move TypedMesh include into class files
