+++ core/src/style/debugStyle.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "style.h"
-#include "gl/typedMesh.h"
 
 namespace Tangram {
 
@@ -9,18 +8,9 @@ class DebugStyle : public Style {
 
 protected:
 
-    struct PosColVertex {
-        // Position Data
-        glm::vec3 pos;
-        // Color Data
-        GLuint abgr;
-    };
-
     virtual void constructVertexLayout() override;
     virtual void constructShaderProgram() override;
     virtual void onBeginBuildTile(Tile& _tile) const override;
-
-    typedef TypedMesh<PosColVertex> Mesh;
 
     virtual VboMesh* newMesh() const override {
         return nullptr;
