+++ core/src/style/debugTextStyle.cpp
@@ -1,10 +1,10 @@
 #include "debugTextStyle.h"
 
 #include "gl/shaderProgram.h"
-#include "gl/vboMesh.h"
 #include "style/material.h"
 #include "text/fontContext.h"
 #include "tile/tile.h"
+#include "tangram.h"
 
 namespace Tangram {
