+++ core/src/style/pointStyle.cpp
@@ -5,8 +5,10 @@
 #include "gl/shaderProgram.h"
 #include "gl/texture.h"
 #include "gl/vertexLayout.h"
+#include "labels/labelMesh.h"
 #include "labels/spriteLabel.h"
 #include "scene/drawRule.h"
+#include "scene/spriteAtlas.h"
 #include "tile/tile.h"
 #include "util/builders.h"
 #include "view/view.h"
@@ -48,6 +50,10 @@ void PointStyle::constructShaderProgram() {
     }
 
     m_shaderProgram->addSourceBlock("defines", defines);
+}
+
+VboMesh* PointStyle::newMesh() const {
+    return new LabelMesh(m_vertexLayout, m_drawMode);
 }
 
 PointStyle::Parameters PointStyle::applyRule(const DrawRule& _rule, const Properties& _props) const {
